requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
pytz==2019.1
gunicorn==19.9.0
whitenoise==4.1.2
django_cleanup==3.2.0
mysqlclient==1.4.2
pillow==6.0.0
coverage==4.5.3
django-heroku==0.3.1
django-ckeditor==5.7.1